Increase versionCode and remove unused variables
This commit increments the `versionCode` to 214.

It also removes the unused `context` variable in `ApkItem.kt` and the unused `isFolder` variable in `FileCard.kt`. The conditional logic for previewing files based on whether it's a folder or a file in `FileCard.kt` has been simplified, as `FilePreviewHelper.Preview` handles both cases.